Upgrade existing scripts/data to use floating objects
