+++ crest/Assets/Crest/Crest-Examples/BoatDev/Data/BoatProbes.prefab
@@ -655,7 +655,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 4572056293069917496}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1adce3e27dc13db40a9debc152108d5b, type: 3}
+  m_Script: {fileID: 11500000, guid: e66dbc64761412d41801763633a21b35, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   _localOffset: {x: 0, y: 0, z: 0}
