--- crest/Assets/Crest/Crest-Examples/BoatDev/Scripts/Boat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 226ac1121586c804bad623882d2e8779
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
