+++ crest/Assets/Crest/Crest/Scripts/Interaction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a58015f9348f394589141c729b4c709
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
